requirements.txt
Django==2.2.1
django-cors-headers==2.5.3
djangorestframework==3.9.3
psycopg2-binary==2.8.2

